requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
Deprecated==1.2.13
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
packaging==21.3
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.6
redis==4.3.4
sqlparse==0.4.3
wrapt==1.14.1
